+++ config/cashier.php
@@ -1,5 +1,6 @@
 <?php
 
+use Laravel\Cashier\Console\WebhookCommand;
 use Laravel\Cashier\Invoices\DompdfInvoiceRenderer;
 
 return [
@@ -46,15 +47,7 @@ return [
     'webhook' => [
         'secret' => env('STRIPE_WEBHOOK_SECRET'),
         'tolerance' => env('STRIPE_WEBHOOK_TOLERANCE', 300),
-        'events' => [
-            'customer.subscription.created',
-            'customer.subscription.updated',
-            'customer.subscription.deleted',
-            'customer.updated',
-            'customer.deleted',
-            'invoice.payment_action_required',
-            'invoice.payment_succeeded',
-        ],
+        'events' => WebhookCommand::DEFAULT_EVENTS,
     ],
 
     /*