+++ src/Concerns/PerformsCharges.php
@@ -148,9 +148,10 @@ trait PerformsCharges
      * @param  int  $quantity
      * @param  array  $sessionOptions
      * @param  array  $customerOptions
+     * @param  array  $productData
      * @return \Laravel\Cashier\Checkout
      */
-    public function checkoutCharge($amount, $name, $quantity = 1, array $sessionOptions = [], array $customerOptions = [])
+    public function checkoutCharge($amount, $name, $quantity = 1, array $sessionOptions = [], array $customerOptions = [], array $productData = [])
     {
         if ($this->isAutomaticTaxEnabled()) {
             throw new LogicException('For now, you cannot use checkout charges in combination with automatic tax calculation.');
@@ -159,9 +160,9 @@ trait PerformsCharges
         return $this->checkout([[
             'price_data' => [
                 'currency' => $this->preferredCurrency(),
-                'product_data' => [
+                'product_data' => array_merge($productData, [
                     'name' => $name,
-                ],
+                ]),
                 'unit_amount' => $amount,
             ],
             'quantity' => $quantity,